* Fix warnings in pgm/transport. * Add missing export pgm_drop_superuser(). * Fix pgm_tsi_hash() to use re-entrant pgm_print_tsi_r(). * Fix return type of pgm_tsi_equal(). * Update all send() return types to gssize. * Update Reed Solomon module to use unsigned types. * Consolidate all sendto()'s to use one function, and to poll() retry upon   non-network related related. * Update transport source with return value notes. * Send functions now only fail on rate control, data is always pushed into   the transmit window. * Fix major typo in NAK pipe processing. * Remove all shadowed variables from transport. * Send stats only update when the sendto() succeeded. * Cleanup non-descript retvals. * Add checks on all sends for actual sent length.

+++ openpgm/pgm/include/pgm/reed_solomon.h
@@ -25,12 +25,12 @@
 
 G_BEGIN_DECLS
 
-int pgm_rs_create (gpointer*, int, int);
+int pgm_rs_create (gpointer*, guint, guint);
 int pgm_rs_destroy (gpointer);
 
-int pgm_rs_encode (gpointer, const void**, int, void*, int);
-int pgm_rs_decode_parity_inline (gpointer, void**, int*, int);
-int pgm_rs_decode_parity_appended (gpointer, void**, int*, int);
+int pgm_rs_encode (gpointer, const void**, guint, void*, gsize);
+int pgm_rs_decode_parity_inline (gpointer, void**, guint*, gsize);
+int pgm_rs_decode_parity_appended (gpointer, void**, guint*, gsize);
 
 #if 0
 int pgm_rs_bch_create (gpointer*, int, int);
